+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -34,7 +34,7 @@ public class ReservationDao {
                 .usingGeneratedKeyColumns("id");
     }
 
-    public List<Reservation> findAllReservations() {
+    public List<Reservation> findAll() {
         String sql = "SELECT r.id AS reservation_id, r.name, r.date, t.id AS time_id, t.start_at AS time_value " +
                 "FROM reservation AS r INNER JOIN reservation_time AS t ON r.time_id = t.id";
         return jdbcTemplate.query(sql, rowMapper);
