+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,37 +1,39 @@
 package roomescape.service;
 
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationTimeDao;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
 import roomescape.dto.ReservationTimeSaveDto;
 import roomescape.entity.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
+
+import java.util.List;
 
 @Service
 public class ReservationTimeService {
 
-    private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationTimeDao reservationTimeDao;
 
     @Autowired
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
     }
 
     public List<ReservationTimeResponseDto> findTimes() {
-        List<ReservationTime> times = reservationTimeRepository.findAll();
+        List<ReservationTime> times = reservationTimeDao.findAll();
         return times.stream()
                 .map(ReservationTimeResponseDto::from)
                 .toList();
     }
 
     public ReservationTimeResponseDto addTime(ReservationTimeRequestDto reservationTimeDto) {
-        ReservationTime reservationTime = reservationTimeRepository.save(ReservationTimeSaveDto.from(reservationTimeDto));
+        long id = reservationTimeDao.save(ReservationTimeSaveDto.from(reservationTimeDto));
+        ReservationTime reservationTime = reservationTimeDao.findById(id);
         return ReservationTimeResponseDto.from(reservationTime);
     }
 
     public void deleteTime(long id) {
-        reservationTimeRepository.delete(id);
+        reservationTimeDao.delete(id);
     }
 }
